Add files via upload
These 3 ipynb programs are a set of progressive steps in how to analyze and deploy a scikit-learn logistic regression model to predict frauds or anomalies from a public credit card dataset, using MLFlow.  The code is based on the book of Alla, S., & Adari, S. (2021). Beginning MLOps with MLFlow: Deploy Models in AWS SageMaker, Google Cloud, and Microsoft Azure. Apress. These programs might be useful as complimentary material for those beginning into the study of principles of MLOps, as well for those reviewing this book as it seems there is not a repository of the code for this book.